third_party/WebKit/LayoutTests/external/wpt/XMLHttpRequest/send-response-event-order-expected.txt

This is a testharness.js-based test.
FAIL XMLHttpRequest: The send() method: event order when synchronous flag is unsetassert_equals: expected "upload.loadstart(0,12,true)" but got "upload.loadstart(0,0,false)"
Harness: the test ran to completion.
